Implementing tests for all the tokens with the new architecture. Also new way to test the tokenization.

+++ core/src/main/java/fr/uge/modules/linking/token/type/TypeDatetime.java
@@ -34,7 +34,7 @@ public class TypeDatetime implements TokenType{
         return new Computation(new TypeDatetime(), ldt1.toString(), ldt2.toString(), fromTime(Math.abs(time), delta));
     }
 
-    private static double fromTime(long time, long delta){
+    private static double fromTime(long time, float delta){
         if(time > delta) return 0;
         else if (time == 0) return 100;
         else {
